API

Parameter dictionaries

CLIMAParameters.AbstractTOMLDict — Type

AbstractTOMLDict{FT <: AbstractFloat}

Abstract parameter dict. Two subtypes:

ParamDict

AliasParamDict

CLIMAParameters.ParamDict — Type

ParamDict(data::Dict, override_dict::Union{Nothing,Dict})

Structure to hold information read-in from TOML file, as well as a parametrization type FT.

Uses the name to search

Fields

data: dictionary representing a default/merged parameter TOML file

override_dict: either a nothing, or a dictionary representing an override parameter TOML file

CLIMAParameters.AliasParamDict — Type

AliasParamDict(data::Dict, override_dict::Union{Nothing,Dict})

Structure to hold information read-in from TOML file, as well as a parametrization type FT.

Uses the alias to search

Fields

data: dictionary representing a default/merged parameter TOML file

override_dict: either a nothing, or a dictionary representing an override parameter TOML file

alias_to_name_map: Alias->name map

File parsing and parameter logging

User facing functions:

CLIMAParameters.create_toml_dict — Function

create_toml_dict(FT;
    override_file,
    default_file,
    dict_type="alias"
)

Creates a ParamDict{FT} struct, by reading and merging upto two TOML files or Julia Dicts with override information taking precedence over default information.

CLIMAParameters.get_parameter_values! — Function

get_parameter_values!(
    pd::AbstractTOMLDict,
    names::Union{String,Vector{String}},
    component::String
)

(Note the !) Gets the parameter values, and logs the component (if given) where parameters are used.

CLIMAParameters.get_parameter_values — Function

get_parameter_values(pd::AbstractTOMLDict, names)

Gets the parameter values only.

CLIMAParameters.float_type — Function

float_type(::AbstractTOMLDict)

The float type from the parameter dict.

CLIMAParameters.log_parameter_information — Function

log_parameter_information(
    pd::AbstractTOMLDict,
    filepath;
    strict::Bool = false
)

Writes the parameter log file at filepath; checks that override parameters are all used.

If strict = true, errors if override parameters are unused.

CLIMAParameters.write_log_file — Function

write_log_file(pd::AbstractTOMLDict, filepath)

Writes a log file of all used parameters of pd at the filepath. This file can be used to rerun the experiment.

CLIMAParameters.merge_toml_files — Function

merge_toml_files(filepaths; override)

Parses and merges all of the given TOML filepaths and returns them as a Dict. This allows a toml_dict to be constructed from multiple TOML files. By default, non-unique TOML entries are not allowed, but this can be changed by setting override = true.