API

Types

CLIMAParameters.AbstractParameterSet — Type

AbstractParameterSet

The top-level super-type parameter set.

CLIMAParameters.AbstractEarthParameterSet — Type

AbstractEarthParameterSet <: AbstractParameterSet

An earth parameter set, specific to planet Earth.

Universal Constants

CLIMAParameters.gas_constant — Function

gas_constant

Universal gas constant (J/mol/K)

CLIMAParameters.light_speed — Function

light_speed

Speed of light in vacuum (m/s)

CLIMAParameters.h_Planck — Function

h_Planck

Planck constant (m^2 kg/s)

CLIMAParameters.k_Boltzmann — Function

k_Boltzmann

Boltzmann constant (m^2 kg/s^2/K)

CLIMAParameters.Stefan — Function

Stefan

Stefan-Boltzmann constant (W/m^2/K^4)

CLIMAParameters.astro_unit — Function

astro_unit

Astronomical unit (m)

CLIMAParameters.avogad — Function

avogad(::Type{FT})

Avogadro's number [molec/mol]

Planet

CLIMAParameters.Planet — Module

Planet

Planetary parameters.

CLIMAParameters.Planet.molmass_dryair — Function

Molecular weight dry air (kg/mol)

CLIMAParameters.Planet.R_d — Function

Gas constant dry air (J/kg/K)

CLIMAParameters.Planet.kappa_d — Function

Adiabatic exponent dry air

CLIMAParameters.Planet.cp_d — Function

Isobaric specific heat dry air

CLIMAParameters.Planet.cv_d — Function

Isochoric specific heat dry air

CLIMAParameters.Planet.ρ_cloud_liq — Function

Density of liquid water (kg/m$^3$)

CLIMAParameters.Planet.ρ_cloud_ice — Function

Density of ice water (kg/m$^3$)

CLIMAParameters.Planet.molmass_water — Function

Molecular weight (kg/mol)

CLIMAParameters.Planet.molmass_ratio — Function

Molar mass ratio dry air/water

CLIMAParameters.Planet.R_v — Function

Gas constant water vapor (J/kg/K)

CLIMAParameters.Planet.cp_v — Function

Isobaric specific heat vapor (J/kg/K)

CLIMAParameters.Planet.cp_l — Function

Isobaric specific heat liquid (J/kg/K)

CLIMAParameters.Planet.cp_i — Function

Isobaric specific heat ice (J/kg/K)

CLIMAParameters.Planet.cv_v — Function

Isochoric specific heat vapor (J/kg/K)

CLIMAParameters.Planet.cv_l — Function

Isochoric specific heat liquid (J/kg/K)

CLIMAParameters.Planet.cv_i — Function

Isochoric specific heat ice (J/kg/K)

CLIMAParameters.Planet.T_freeze — Function

Freezing point temperature (K)

CLIMAParameters.Planet.T_min — Function

Minimum temperature guess in saturation adjustment (K)

CLIMAParameters.Planet.T_max — Function

Maximum temperature guess in saturation adjustment (K)

CLIMAParameters.Planet.T_icenuc — Function

Homogeneous nucleation temperature (K)

CLIMAParameters.Planet.pow_icenuc — Function

The assumed phase partition function power (-)

CLIMAParameters.Planet.T_triple — Function

Triple point temperature (K)

CLIMAParameters.Planet.T_0 — Function

Reference temperature (K)

CLIMAParameters.Planet.LH_v0 — Function

Latent heat vaporization at $T_0$ (J/kg)

CLIMAParameters.Planet.LH_s0 — Function

Latent heat sublimation at $T_0$ (J/kg)

CLIMAParameters.Planet.LH_f0 — Function

Latent heat of fusion at $T_0$ (J/kg)

CLIMAParameters.Planet.e_int_v0 — Function

Specific internal energy of vapor at $T_0$ (J/kg)

CLIMAParameters.Planet.e_int_i0 — Function

Specific internal energy of ice at $T_0$ (J/kg)

CLIMAParameters.Planet.press_triple — Function

Triple point vapor pressure (Pa)

CLIMAParameters.Planet.surface_tension_coeff — Function

Surface tension coefficient of water (J/m2)

CLIMAParameters.Planet.entropy_water_vapor — Function

Entropy of water vapor J / (kg K)

CLIMAParameters.Planet.entropy_dry_air — Function

Entropy of dry air J / (kg K)

CLIMAParameters.Planet.entropy_reference_temperature — Function

Entropy reference temperature (K)

CLIMAParameters.Planet.ρ_ocean — Function

Reference density sea water (kg/m$^3$)

CLIMAParameters.Planet.cp_ocean — Function

Specific heat sea water (J/kg/K)

CLIMAParameters.Planet.planet_radius — Function

Mean planetary radius (m)

CLIMAParameters.Planet.day — Function

Length of day (s)

CLIMAParameters.Planet.Omega — Function

Ang. velocity planetary rotation (1/s)

CLIMAParameters.Planet.grav — Function

Gravitational acceleration (m/s$^2$)

CLIMAParameters.Planet.year_anom — Function

Length of anomalistic year (s)

CLIMAParameters.Planet.orbit_semimaj — Function

ngth of semimajor orbital axis (m)

CLIMAParameters.Planet.tot_solar_irrad — Function

Total solar irradiance (W/m$^2$)

CLIMAParameters.Planet.epoch — Function

Time of epoch (J2000) (s)

CLIMAParameters.Planet.mean_anom_epoch — Function

Mean anomaly at the epoch (radians)

CLIMAParameters.Planet.obliq_epoch — Function

Orbital obliquity at the epoch (radians)

CLIMAParameters.Planet.lon_perihelion_epoch — Function

Longitude of perihelion at the epoch (radians), measured relative to vernal equinox (i.e., the longitude of perihelion is the angle subtended at the Sun by the orbital arc from vernal equinox to perihelion).

CLIMAParameters.Planet.eccentricity_epoch — Function

Orbital eccentricity at the epoch

CLIMAParameters.Planet.lon_perihelion — Function

Longitude of perihelion (radians), measured relative to vernal equinox. The calculation of the mean anomaly is formulated such that the vernal equinox is fixed in the calendar. However, this requires tracking both the current longitude of perihelion as well as the reference longitude of perihelion at the reference time (epoch).

CLIMAParameters.Planet.MSLP — Function

Mean sea level pressure (Pa)

CLIMAParameters.Planet.T_surf_ref — Function

Mean surface temperature (K) in reference state

CLIMAParameters.Planet.T_min_ref — Function

Minimum temperature (K) in reference state

Common

Sub-grid scale

CLIMAParameters.SubgridScale — Module

SubgridScale

Sub-grid scale parameters, shared by multiple models including the atmosphere, ocean and surface fluxes.

CLIMAParameters.SubgridScale.von_karman_const — Function

Von Karman constant (dimensionless)

Surface fluxes and universal functions

CLIMAParameters.SurfaceFluxes — Module

SurfaceFluxes

Surface fluxes parameters, shared by multiple models including the atmosphere, ocean.

CLIMAParameters.SurfaceFluxes.UniversalFunctions — Module

UniversalFunctions

Universal functions for the surface fluxes module.

CLIMAParameters.SurfaceFluxes.UniversalFunctions.Pr_0_Businger — Function

Prandtl number at neutral stratification

CLIMAParameters.SurfaceFluxes.UniversalFunctions.a_m_Businger — Function

coefficient for momentum

CLIMAParameters.SurfaceFluxes.UniversalFunctions.a_h_Businger — Function

coefficient for heat

CLIMAParameters.SurfaceFluxes.UniversalFunctions.Pr_0_Gryanik — Function

neutral-limit turbulent Prandtl number

CLIMAParameters.SurfaceFluxes.UniversalFunctions.a_m_Gryanik — Function

empirical coefficient (a) for momentum

CLIMAParameters.SurfaceFluxes.UniversalFunctions.a_h_Gryanik — Function

empirical coefficient (a) for heat

CLIMAParameters.SurfaceFluxes.UniversalFunctions.b_m_Gryanik — Function

empirical coefficient (b) for momentum

CLIMAParameters.SurfaceFluxes.UniversalFunctions.b_h_Gryanik — Function

empirical coefficient (b) for heat

CLIMAParameters.SurfaceFluxes.UniversalFunctions.Pr_0_Grachev — Function

Prandtl number at neutral stratification

CLIMAParameters.SurfaceFluxes.UniversalFunctions.a_m_Grachev — Function

empirical coefficient (a) for momentum

CLIMAParameters.SurfaceFluxes.UniversalFunctions.a_h_Grachev — Function

empirical coefficient (a) for heat

CLIMAParameters.SurfaceFluxes.UniversalFunctions.b_m_Grachev — Function

empirical coefficient (b) for momentum

CLIMAParameters.SurfaceFluxes.UniversalFunctions.b_h_Grachev — Function

empirical coefficient (b) for heat

CLIMAParameters.SurfaceFluxes.UniversalFunctions.c_h_Grachev — Function

empirical coefficient (c) for heat

Atmos

CLIMAParameters.Atmos — Module

Atmos

Atmospheric parameters.

Sub-grid scale

CLIMAParameters.Atmos.SubgridScale.C_smag — Function

Smagorinsky Coefficient (dimensionless)

CLIMAParameters.Atmos.SubgridScale.C_drag — Function

Drag coefficient (dimensionless)

CLIMAParameters.Atmos.SubgridScale.inv_Pr_turb — Function

Turbulent Prandtl Number (dimensionless)

CLIMAParameters.Atmos.SubgridScale.Prandtl_air — Function

Molecular Prandtl Number, dry air (dimensionless)

CLIMAParameters.Atmos.SubgridScale.c_a_KASM — Function

$c_{a}$ KASM, 2006 (dimensionless)

CLIMAParameters.Atmos.SubgridScale.c_e1_KASM — Function

$c_{e1}$ KASM, 2006 (dimensionless)

CLIMAParameters.Atmos.SubgridScale.c_e2_KASM — Function

$c_{e2}$ KASM, 2006 (dimensionless)

CLIMAParameters.Atmos.SubgridScale.c_1_KASM — Function

$c_{1}$ KASM, 2006 (dimensionless)

CLIMAParameters.Atmos.SubgridScale.c_2_KASM — Function

$c_{2}$ KASM, 2006 (dimensionless)

CLIMAParameters.Atmos.SubgridScale.c_3_KASM — Function

$c_{3}$ KASM, 2006 (dimensionless)

EDMF

CLIMAParameters.Atmos.EDMF.c_λ — Function

Entrainment TKE scale (dimensionless)

CLIMAParameters.Atmos.EDMF.c_ε — Function

Entrainment factor (dimensionless)

CLIMAParameters.Atmos.EDMF.c_δ — Function

Detrainment factor (dimensionless)

CLIMAParameters.Atmos.EDMF.c_γ — Function

Turbulent Entrainment factor (dimensionless)

CLIMAParameters.Atmos.EDMF.β — Function

Detrainment RH power (dimensionless)

CLIMAParameters.Atmos.EDMF.μ_0 — Function

Logistic function scale (s$^{-1}$)

CLIMAParameters.Atmos.EDMF.χ — Function

Updraft mixing fraction (dimensionless)

CLIMAParameters.Atmos.EDMF.w_min — Function

Minimum updraft velocity (m/s)

CLIMAParameters.Atmos.EDMF.lim_ϵ — Function

Exponential area limiter scale (dimensionless)

CLIMAParameters.Atmos.EDMF.lim_amp — Function

Exponential area limiter amplitude(dimensionless)

CLIMAParameters.Atmos.EDMF.a_min — Function

Minimum area fraction for any subdomain (dimensionless)

CLIMAParameters.Atmos.EDMF.a_surf — Function

Updraft surface area fraction (dimensionless)

CLIMAParameters.Atmos.EDMF.κ_star² — Function

Square ratio of rms turbulent velocity to friction velocity (dimensionless)

CLIMAParameters.Atmos.EDMF.ψϕ_stab — Function

Surface covariance stability coefficient (dimensionless)

CLIMAParameters.Atmos.EDMF.α_d — Function

Pressure drag factor (dimensionless)

CLIMAParameters.Atmos.EDMF.α_a — Function

Advective pressure factor (dimensionless

CLIMAParameters.Atmos.EDMF.α_b — Function

Buoyancy pressure factor (dimensionless)

CLIMAParameters.Atmos.EDMF.H_up_min — Function

Minimum updraft height for closures (m)

CLIMAParameters.Atmos.EDMF.c_d — Function

Environmental dissipation factor (dimensionless)

CLIMAParameters.Atmos.EDMF.c_m — Function

Eddy viscosity factor (dimensionless)

CLIMAParameters.Atmos.EDMF.c_b — Function

Static stability factor (dimensionless)

CLIMAParameters.Atmos.EDMF.a1 — Function

Empirical stability function coefficient (dimensionless)

CLIMAParameters.Atmos.EDMF.a2 — Function

Empirical stability function coefficient (dimensionless)

CLIMAParameters.Atmos.EDMF.ω_pr — Function

Prandtl number empirical coefficient (dimensionless)

CLIMAParameters.Atmos.EDMF.Pr_n — Function

Neutral Prandtl number (dimensionless)

CLIMAParameters.Atmos.EDMF.Ri_c — Function

Critical Richardson number (dimensionless)

CLIMAParameters.Atmos.EDMF.smin_ub — Function

Smooth minimum's fractional upper bound (dimensionless)

CLIMAParameters.Atmos.EDMF.smin_rm — Function

Smooth minimum's regularization length minimum (m)

Microphysics_ne

CLIMAParameters.Atmos.Microphysics_ne.τ_cond_evap — Function

condensation/evaporation timescale (s)

CLIMAParameters.Atmos.Microphysics_ne.τ_sub_dep — Function

sublimation/deposition timescale (s)

Microphysics_0M

CLIMAParameters.Atmos.Microphysics_0M.τ_precip — Function

precipitation removal timescale (s)

CLIMAParameters.Atmos.Microphysics_0M.qc_0 — Function

precipitation removal threshold expressed in condensate specific humidity (kg/kg)

CLIMAParameters.Atmos.Microphysics_0M.S_0 — Function

precipitation removal threshold expressed in supersaturation (dimensionless)

Microphysics_1M

Please see the microphysics documentation for an explanation of the default values.

CLIMAParameters.Atmos.Microphysics_1M.C_drag — Function

drag coefficient for rain drops (-)

CLIMAParameters.Atmos.Microphysics_1M.K_therm — Function

thermal conductivity of air (J/m/s/K)

CLIMAParameters.Atmos.Microphysics_1M.D_vapor — Function

diffusivity of water vapor (m$^2$/s)

CLIMAParameters.Atmos.Microphysics_1M.ν_air — Function

kinematic viscosity of air (m$^2$/s)

CLIMAParameters.Atmos.Microphysics_1M.N_Sc — Function

Schmidt number (-)

CLIMAParameters.Atmos.Microphysics_1M.r_ice_snow — Function

threshold between ice and snow (m)

CLIMAParameters.Atmos.Microphysics_1M.n0_ice — Function

Marshall-Palmer distribution n_0 coeff for rain or snow (1/m$^4$)

CLIMAParameters.Atmos.Microphysics_1M.r0_ice — Function

assumed length scale for water drop or ice crystal (m)

CLIMAParameters.Atmos.Microphysics_1M.me_ice — Function

exponent in mass, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics_1M.m0_ice — Function

coefficient in mass, radius/r0, (kg)

CLIMAParameters.Atmos.Microphysics_1M.χm_ice — Function

coefficient in mass, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics_1M.Δm_ice — Function

exponent in mass, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics_1M.q_liq_threshold — Function

rain autoconversion threshold ∈(0.5, 1) * 1e-3 (-)

CLIMAParameters.Atmos.Microphysics_1M.τ_acnv_rai — Function

rain autoconversion timescale ∈(1e3, 1e4) (s)

CLIMAParameters.Atmos.Microphysics_1M.a_vent_rai — Function

ventilation factor coefficient for rain (-)

CLIMAParameters.Atmos.Microphysics_1M.b_vent_rai — Function

ventilation factor coefficient for rain (-)

CLIMAParameters.Atmos.Microphysics_1M.n0_rai — Function

Marshall-Palmer distribution n_0 coeff for rain or snow (1/m$^4$)

CLIMAParameters.Atmos.Microphysics_1M.r0_rai — Function

assumed length scale for water drop or ice crystal (m)

CLIMAParameters.Atmos.Microphysics_1M.me_rai — Function

exponent in mass, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics_1M.ae_rai — Function

exponent in area, radius/r0, relation

CLIMAParameters.Atmos.Microphysics_1M.ve_rai — Function

exponent in velocity, radius/r0,

CLIMAParameters.Atmos.Microphysics_1M.m0_rai — Function

coefficient in mass, radius/r0, (kg)

CLIMAParameters.Atmos.Microphysics_1M.a0_rai — Function

coefficient in area, radius/r0, relation (m$^2$)

CLIMAParameters.Atmos.Microphysics_1M.χm_rai — Function

coefficient in mass, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics_1M.Δm_rai — Function

exponent in mass, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics_1M.χa_rai — Function

coefficient in area, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics_1M.Δa_rai — Function

exponent in area, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics_1M.χv_rai — Function

coefficient in velocity, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics_1M.Δv_rai — Function

exponent in velocity, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics_1M.q_ice_threshold — Function

snow autoconversion threshold (-)

CLIMAParameters.Atmos.Microphysics_1M.τ_acnv_sno — Function

snow autoconversion timescale (s)

CLIMAParameters.Atmos.Microphysics_1M.a_vent_sno — Function

ventilation factor coefficient for snow (-)

CLIMAParameters.Atmos.Microphysics_1M.b_vent_sno — Function

ventilation factor coefficient for snow (-)

CLIMAParameters.Atmos.Microphysics_1M.μ_sno — Function

coefficient to compute Marshall-Palmer distribution coefficient n_0(ρq_snow/ρ0) for snow (1/m$^4$)

CLIMAParameters.Atmos.Microphysics_1M.ν_sno — Function

coefficient to compute Marshall-Palmer distribution coefficient n_0(ρq_snow/ρ0) for snow (-)

CLIMAParameters.Atmos.Microphysics_1M.r0_sno — Function

assumed length scale for water drop or ice crystal (m)

CLIMAParameters.Atmos.Microphysics_1M.me_sno — Function

exponent in mass, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics_1M.ae_sno — Function

exponent in area, radius/r0, relation

CLIMAParameters.Atmos.Microphysics_1M.ve_sno — Function

exponent in velocity, radius/r0,

CLIMAParameters.Atmos.Microphysics_1M.m0_sno — Function

coefficient in mass, radius/r0, (kg)

CLIMAParameters.Atmos.Microphysics_1M.a0_sno — Function

coefficient in area, radius/r0, relation (m$^2$)

CLIMAParameters.Atmos.Microphysics_1M.v0_sno — Function

coefficient in velocity, radius/r0, (m/s)

CLIMAParameters.Atmos.Microphysics_1M.χm_sno — Function

coefficient in mass, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics_1M.Δm_sno — Function

exponent in mass, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics_1M.χa_sno — Function

coefficient in area, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics_1M.Δa_sno — Function

exponent in area, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics_1M.χv_sno — Function

coefficient in velocity, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics_1M.Δv_sno — Function

exponent in velocity, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics_1M.microph_scaling — Function

rain evaporation scaling factor for coupling with EDMF (-)

CLIMAParameters.Atmos.Microphysics_1M.microph_scaling_dep_sub — Function

snow deposition and sublimation scaling factor for coupling with EDMF (-)

CLIMAParameters.Atmos.Microphysics_1M.microph_scaling_melt — Function

snow melting scaling factor for coupling with EDMF (-)

CLIMAParameters.Atmos.Microphysics_1M.E_liq_rai — Function

collision efficiency (-)

CLIMAParameters.Atmos.Microphysics_1M.E_liq_sno — Function

collision efficiency (-)

CLIMAParameters.Atmos.Microphysics_1M.E_ice_rai — Function

collision efficiency (-)

CLIMAParameters.Atmos.Microphysics_1M.E_ice_sno — Function

collision efficiency (-)

CLIMAParameters.Atmos.Microphysics_1M.E_rai_sno — Function

collision efficiency (-)