API

Types

CLIMAParameters.AbstractParameterSet — Type

AbstractParameterSet

The top-level super-type parameter set.

CLIMAParameters.AbstractEarthParameterSet — Type

AbstractEarthParameterSet <: AbstractParameterSet

An earth parameter set, specific to planet Earth.

CLIMAParameters.AbstractMicrophysicsParameterSet — Type

AbstractMicrophysicsParameterSet <: AbstractParameterSet

A set of parameters for cloud and precipitation microphysics parameterization.

CLIMAParameters.AbstractCloudParameterSet — Type

AbstractCloudParameterSet  <: AbstractMicrophysicsParameterSet

A set of parameters for cloud microphysics parameterization.

CLIMAParameters.AbstractPrecipParameterSet — Type

AbstractPrecipParameterSet <: AbstractMicrophysicsParameterSet

A set of parameters for precipitation microphysics parameterization.

CLIMAParameters.AbstractLiquidParameterSet — Type

AbstractLiquidParameterSet <: AbstractCloudParameterSet

A set of parameters for cloud liquid water microphysics parameterization.

CLIMAParameters.AbstractIceParameterSet — Type

AbstractIceParameterSet    <: AbstractCloudParameterSet

A set of parameters for cloud ice microphysics parameterization.

CLIMAParameters.AbstractRainParameterSet — Type

AbstractRainParameterSet   <: AbstractPrecipParameterSet

A set of parameters for rain microphysics parameterization.

CLIMAParameters.AbstractSnowParameterSet — Type

AbstractSnowParameterSet   <: AbstractPrecipParameterSet

A set of parameters for snow microphysics parameterization.

Universal Constants

CLIMAParameters.gas_constant — Function

gas_constant

Universal gas constant (J/mol/K)

CLIMAParameters.light_speed — Function

light_speed

Speed of light in vacuum (m/s)

CLIMAParameters.h_Planck — Function

h_Planck

Planck constant (m^2 kg/s)

CLIMAParameters.k_Boltzmann — Function

k_Boltzmann

Boltzmann constant (m^2 kg/s^2/K)

CLIMAParameters.Stefan — Function

Stefan

Stefan-Boltzmann constant (W/m^2/K^4)

CLIMAParameters.astro_unit — Function

astro_unit

Astronomical unit (m)

CLIMAParameters.avogad — Function

avogad(::Type{FT})

Avogadro's number [molec/mol]

Planet

CLIMAParameters.Planet — Module

Planet

Planetary parameters.

CLIMAParameters.Planet.molmass_dryair — Function

Molecular weight dry air (kg/mol)

CLIMAParameters.Planet.R_d — Function

Gas constant dry air (J/kg/K)

CLIMAParameters.Planet.kappa_d — Function

Adiabatic exponent dry air

CLIMAParameters.Planet.cp_d — Function

Isobaric specific heat dry air

CLIMAParameters.Planet.cv_d — Function

Isochoric specific heat dry air

CLIMAParameters.Planet.ρ_cloud_liq — Function

Density of liquid water (kg/m$^3$)

CLIMAParameters.Planet.ρ_cloud_ice — Function

Density of ice water (kg/m$^3$)

CLIMAParameters.Planet.molmass_water — Function

Molecular weight (kg/mol)

CLIMAParameters.Planet.molmass_ratio — Function

Molar mass ratio dry air/water

CLIMAParameters.Planet.R_v — Function

Gas constant water vapor (J/kg/K)

CLIMAParameters.Planet.cp_v — Function

Isobaric specific heat vapor (J/kg/K)

CLIMAParameters.Planet.cp_l — Function

Isobaric specific heat liquid (J/kg/K)

CLIMAParameters.Planet.cp_i — Function

Isobaric specific heat ice (J/kg/K)

CLIMAParameters.Planet.cv_v — Function

Isochoric specific heat vapor (J/kg/K)

CLIMAParameters.Planet.cv_l — Function

Isochoric specific heat liquid (J/kg/K)

CLIMAParameters.Planet.cv_i — Function

Isochoric specific heat ice (J/kg/K)

CLIMAParameters.Planet.T_freeze — Function

Freezing point temperature (K)

CLIMAParameters.Planet.T_min — Function

Minimum temperature guess in saturation adjustment (K)

CLIMAParameters.Planet.T_max — Function

Maximum temperature guess in saturation adjustment (K)

CLIMAParameters.Planet.T_icenuc — Function

Homogeneous nucleation temperature (K)

CLIMAParameters.Planet.T_triple — Function

Triple point temperature (K)

CLIMAParameters.Planet.T_0 — Function

Reference temperature (K)

CLIMAParameters.Planet.LH_v0 — Function

Latent heat vaporization at $T_0$ (J/kg)

CLIMAParameters.Planet.LH_s0 — Function

Latent heat sublimation at $T_0$ (J/kg)

CLIMAParameters.Planet.LH_f0 — Function

Latent heat of fusion at $T_0$ (J/kg)

CLIMAParameters.Planet.e_int_v0 — Function

Specific internal energy of vapor at $T_0$ (J/kg)

CLIMAParameters.Planet.e_int_i0 — Function

Specific internal energy of ice at $T_0$ (J/kg)

CLIMAParameters.Planet.press_triple — Function

Triple point vapor pressure (Pa)

CLIMAParameters.Planet.ρ_ocean — Function

Reference density sea water (kg/m$^3$)

CLIMAParameters.Planet.cp_ocean — Function

Specific heat sea water (J/kg/K)

CLIMAParameters.Planet.planet_radius — Function

Mean planetary radius (m)

CLIMAParameters.Planet.day — Function

Length of day (s)

CLIMAParameters.Planet.Omega — Function

Ang. velocity planetary rotation (1/s)

CLIMAParameters.Planet.grav — Function

Gravitational acceleration (m/s$^2$)

CLIMAParameters.Planet.year_anom — Function

Length of anomalistic year (s)

CLIMAParameters.Planet.orbit_semimaj — Function

ngth of semimajor orbital axis (m)

CLIMAParameters.Planet.TSI — Function

Total solar irradiance (W/m$^2$)

CLIMAParameters.Planet.MSLP — Function

Mean sea level pressure (Pa)

CLIMAParameters.Planet.T_surf_ref — Function

Mean surface temperature (K) in reference state

CLIMAParameters.Planet.T_min_ref — Function

Minimum temperature (K) in reference state

Common

Sub-grid scale

CLIMAParameters.SubgridScale — Module

SubgridScale

Sub-grid scale parameters, shared by multiple models including the atmosphere, ocean and surface fluxes.

CLIMAParameters.SubgridScale.von_karman_const — Function

Von Karman constant (dimensionless)

Atmos

CLIMAParameters.Atmos — Module

Atmos

Atmospheric parameters.

Sub-grid scale

CLIMAParameters.Atmos.SubgridScale.C_smag — Function

Smagorinsky Coefficient (dimensionless)

CLIMAParameters.Atmos.SubgridScale.C_drag — Function

Drag coefficient (dimensionless)

CLIMAParameters.Atmos.SubgridScale.inv_Pr_turb — Function

Turbulent Prandtl Number (dimensionless)

CLIMAParameters.Atmos.SubgridScale.Prandtl_air — Function

Molecular Prandtl Number, dry air (dimensionless)

CLIMAParameters.Atmos.SubgridScale.c_a_KASM — Function

$c_{a}$ KASM, 2006 (dimensionless)

CLIMAParameters.Atmos.SubgridScale.c_e1_KASM — Function

$c_{e1}$ KASM, 2006 (dimensionless)

CLIMAParameters.Atmos.SubgridScale.c_e2_KASM — Function

$c_{e2}$ KASM, 2006 (dimensionless)

CLIMAParameters.Atmos.SubgridScale.c_1_KASM — Function

$c_{1}$ KASM, 2006 (dimensionless)

CLIMAParameters.Atmos.SubgridScale.c_2_KASM — Function

$c_{2}$ KASM, 2006 (dimensionless)

CLIMAParameters.Atmos.SubgridScale.c_3_KASM — Function

$c_{3}$ KASM, 2006 (dimensionless)

Microphysics_0M

CLIMAParameters.Atmos.Microphysics_0M.τ_precip — Function

precipitation removal timescale (s)

CLIMAParameters.Atmos.Microphysics_0M.qc_0 — Function

precipitation removal threshold expressed in condensate specific humidity (kg/kg)

CLIMAParameters.Atmos.Microphysics_0M.S_0 — Function

precipitation removal threshold expressed in supersaturation (dimensionless)

Microphysics

Please see the microphysics documentation for an explanation of the default values.

CLIMAParameters.Atmos.Microphysics.n0 — Function

Marshall-Palmer distribution n_0 coeff for rain or snow (1/m$^4$)

CLIMAParameters.Atmos.Microphysics.μ_sno — Function

coefficient to compute Marshall-Palmer distribution coefficient n_0(ρq_snow/ρ0) for snow (1/m$^4$)

CLIMAParameters.Atmos.Microphysics.ν_sno — Function

coefficient to compute Marshall-Palmer distribution coefficient n_0(ρq_snow/ρ0) for snow (-)

CLIMAParameters.Atmos.Microphysics.r0 — Function

assumed length scale for water drop or ice crystal (m)

CLIMAParameters.Atmos.Microphysics.m0 — Function

coefficient in mass, radius/r0, (kg)

CLIMAParameters.Atmos.Microphysics.χm — Function

coefficient in mass, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics.me — Function

exponent in mass, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics.Δm — Function

exponent in mass, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics.a0 — Function

coefficient in area, radius/r0, relation (m$^2$)

CLIMAParameters.Atmos.Microphysics.χa — Function

coefficient in area, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics.ae — Function

exponent in area, radius/r0, relation

CLIMAParameters.Atmos.Microphysics.Δa — Function

exponent in area, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics.v0 — Function

coefficient in velocity, radius/r0, (m/s)

CLIMAParameters.Atmos.Microphysics.χv — Function

coefficient in velocity, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics.ve — Function

exponent in velocity, radius/r0,

CLIMAParameters.Atmos.Microphysics.Δv — Function

exponent in velocity, radius/r0, (-)

CLIMAParameters.Atmos.Microphysics.C_drag — Function

drag coefficient for rain drops (-)

CLIMAParameters.Atmos.Microphysics.τ_cond_evap — Function

condensation/evaporation timescale (s)

CLIMAParameters.Atmos.Microphysics.τ_sub_dep — Function

sublimation/deposition timescale (s)

CLIMAParameters.Atmos.Microphysics.q_liq_threshold — Function

autoconversion threshold ∈(0.5, 1) * 1e-3 (-)

CLIMAParameters.Atmos.Microphysics.τ_acnv — Function

autoconversion timescale ∈(1e3, 1e4) (s)

CLIMAParameters.Atmos.Microphysics.r_ice_snow — Function

threshold between ice and snow (m)

CLIMAParameters.Atmos.Microphysics.E — Function

collision efficiency (-)

CLIMAParameters.Atmos.Microphysics.a_vent — Function

ventilation factor coefficient for rain or snow (-)

CLIMAParameters.Atmos.Microphysics.b_vent — Function

ventilation factor coefficient for rain or snow (-)

CLIMAParameters.Atmos.Microphysics.K_therm — Function

thermal conductivity of air (J/m/s/K)

CLIMAParameters.Atmos.Microphysics.D_vapor — Function

diffusivity of water vapor (m$^2$/s)

CLIMAParameters.Atmos.Microphysics.ν_air — Function

kinematic viscosity of air (m$^2$/s)

CLIMAParameters.Atmos.Microphysics.N_Sc — Function

Schmidt number (-)